services/outreach-agent/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9

# LLM
anthropic==0.39.0
openai==1.54.0

# Agent framework
langgraph==0.2.53
langchain-core==0.3.21
langchain-anthropic==0.3.0

# Gmail API
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Scheduling
APScheduler==3.10.4

# Data processing (CSV/XLSX upload)
pandas==2.2.3
openpyxl==3.1.5

# Utilities
httpx==0.27.2
python-multipart==0.0.9
email-validator==2.2.0
tenacity==9.0.0
